+++ scripts/variables.sh
@@ -25,6 +25,7 @@ restore_processes=""
 restore_process_strategy_option="@resurrect-strategy-"
 
 inline_strategy_token="->"
+inline_strategy_arguments_token="*"
 
 save_command_strategy_option="@resurrect-save-command-strategy"
 default_save_command_strategy="ps"